+++ src/test/java/test/org/fugerit/java/nativehelper/maven/TestMojoGenerate.java
@@ -1,12 +1,17 @@
 package test.org.fugerit.java.nativehelper.maven;
 
+import org.apache.maven.artifact.DependencyResolutionRequiredException;
+import org.apache.maven.model.Build;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.project.MavenProject;
 import org.fugerit.java.nativehelper.maven.MojoGenerate;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
 
 class TestMojoGenerate {
 
@@ -25,6 +30,23 @@ class TestMojoGenerate {
         MojoGenerate mojo = new MojoGenerate();
         File file = new File( "target/override-test/native-image/reflect-config.json" );
         file.delete();
+        // maven project mock
+        mojo.setProject( new MavenProject() {
+            @Override
+            public Build getBuild() {
+                return new Build() {
+                    @Override
+                    public String getOutputDirectory() {
+                        return "target";
+                    }
+                };
+            }
+
+            @Override
+            public List<String> getCompileClasspathElements() throws DependencyResolutionRequiredException {
+                return Arrays.asList( new File( this.getBuild().getOutputDirectory(),"classes" ).getAbsolutePath() );
+            }
+        } );
         mojo.setNativeHelperConfigPath( "src/test/resources/tool/config/native-helper-config.yaml" );
         mojo.setReflectConfigJsonOutputPath( file.getCanonicalPath() );
         mojo.execute();
@@ -43,6 +65,15 @@ class TestMojoGenerate {
         MojoGenerate mojo = new MojoGenerate();
         mojo.setNativeHelperConfigPath( "src/test/resources/tool/config/native-helper-config-not-exists.yaml" );
         Assertions.assertThrows( MojoExecutionException.class, () -> mojo.execute() );
+    }
+
+    @Test
+    void testFail3() throws MojoExecutionException, IOException {
+        MojoGenerate mojo = new MojoGenerate();
+        mojo.setNativeHelperConfigPath( "src/test/resources/tool/config/native-helper-config-fail.yaml" );
+        mojo.setWarnOnError( true );
+        mojo.execute();
+        Assertions.assertNotNull( mojo );
     }
 
 }