+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2024-04-01
+
+### Added
+
+- now plugin can read classes generated in current maven project
+
+### Changed
+
+- [native-helper-graalvm](https://github.com/fugerit-org/native-helper-graalvm/) version set to 1.3.1
+
 ## [0.1.0] - 2024-04-01
 
 ### Added